+++ crates/krilla/src/surface.rs
@@ -622,139 +622,3 @@ fn naive_shape(
 
     glyphs
 }
-
-#[cfg(test)]
-mod tests {
-    use krilla_macros::snapshot;
-    use pdf_writer::types::BlendMode;
-    use tiny_skia_path::{Point, Size, Transform};
-
-    use crate::font::Font;
-    use crate::mask::MaskType;
-    use crate::path::Fill;
-    use crate::surface::Stroke;
-    #[cfg(feature = "simple-text")]
-    use crate::surface::TextDirection;
-    use crate::tests::{
-        basic_mask, blue_fill, cmyk_fill, gray_fill, green_fill, load_png_image, rect_to_path,
-        red_fill, NOTO_SANS,
-    };
-    use crate::Page;
-
-    #[snapshot(single_page)]
-    fn stream_path_single_with_rgb(page: &mut Page) {
-        let mut surface = page.surface();
-        let path = rect_to_path(20.0, 20.0, 180.0, 180.0);
-        let fill = red_fill(1.0);
-        surface.fill_path(&path, fill);
-    }
-
-    #[snapshot(single_page)]
-    fn stream_path_single_with_luma(page: &mut Page) {
-        let mut surface = page.surface();
-        let path = rect_to_path(20.0, 20.0, 180.0, 180.0);
-        let fill = gray_fill(1.0);
-        surface.fill_path(&path, fill);
-    }
-
-    #[snapshot(single_page)]
-    fn stream_path_single_with_rgb_and_opacity(page: &mut Page) {
-        let mut surface = page.surface();
-        let path = rect_to_path(20.0, 20.0, 180.0, 180.0);
-        let fill = red_fill(0.5);
-        surface.fill_path(&path, fill);
-    }
-
-    #[snapshot(single_page)]
-    fn stream_path_single_with_cmyk(page: &mut Page) {
-        let mut surface = page.surface();
-        let path = rect_to_path(20.0, 20.0, 180.0, 180.0);
-        let fill = cmyk_fill(1.0);
-        surface.fill_path(&path, fill);
-    }
-
-    #[snapshot(single_page, settings_2)]
-    fn stream_resource_cache(page: &mut Page) {
-        let mut surface = page.surface();
-        let path1 = rect_to_path(0.0, 0.0, 100.0, 100.0);
-        let path2 = rect_to_path(50.0, 50.0, 150.0, 150.0);
-        let path3 = rect_to_path(100.0, 100.0, 200.0, 200.0);
-
-        surface.fill_path(&path1, green_fill(1.0));
-        surface.fill_path(&path2, red_fill(1.0));
-        surface.fill_path(&path3, blue_fill(1.0));
-    }
-
-    #[snapshot(single_page)]
-    fn stream_nested_transforms(page: &mut Page) {
-        let mut surface = page.surface();
-        let path1 = rect_to_path(0.0, 0.0, 100.0, 100.0);
-
-        surface.push_transform(&Transform::from_translate(50.0, 50.0));
-        surface.fill_path(&path1, green_fill(1.0));
-        surface.push_transform(&Transform::from_translate(100.0, 100.0));
-        surface.fill_path(&path1, red_fill(1.0));
-
-        surface.pop();
-        surface.pop();
-    }
-
-    #[snapshot(single_page)]
-    fn stream_reused_graphics_state(page: &mut Page) {
-        let mut surface = page.surface();
-        let path1 = rect_to_path(0.0, 0.0, 100.0, 100.0);
-        surface.fill_path(&path1, green_fill(0.5));
-        surface.push_blend_mode(BlendMode::ColorBurn);
-        surface.fill_path(&path1, green_fill(0.5));
-        surface.pop();
-        surface.fill_path(&path1, green_fill(0.5));
-    }
-
-    #[snapshot(single_page)]
-    fn stream_fill_text(page: &mut Page) {
-        let mut surface = page.surface();
-        surface.fill_text(
-            Point::from_xy(0.0, 50.0),
-            Fill::default(),
-            Font::new(NOTO_SANS.clone(), 0, true).unwrap(),
-            16.0,
-            &[],
-            "hi there",
-            false,
-            TextDirection::Auto,
-        );
-    }
-
-    #[snapshot(single_page)]
-    fn stream_stroke_text(page: &mut Page) {
-        let mut surface = page.surface();
-        surface.stroke_text(
-            Point::from_xy(0.0, 50.0),
-            Stroke::default(),
-            Font::new(NOTO_SANS.clone(), 0, true).unwrap(),
-            16.0,
-            &[],
-            "hi there",
-            false,
-            TextDirection::Auto,
-        );
-    }
-
-    #[snapshot(single_page)]
-    fn stream_image(page: &mut Page) {
-        let mut surface = page.surface();
-        let image = load_png_image("rgb8.png");
-        let size = Size::from_wh(image.size().0 as f32, image.size().1 as f32).unwrap();
-        surface.draw_image(image, size);
-    }
-
-    #[snapshot(single_page)]
-    fn stream_mask(page: &mut Page) {
-        let mut surface = page.surface();
-        let mask = basic_mask(&mut surface, MaskType::Alpha);
-        surface.push_mask(mask);
-        let path = rect_to_path(0.0, 0.0, 100.0, 100.0);
-        surface.fill_path(&path, green_fill(0.5));
-        surface.pop();
-    }
-}
